Apply nullToValue transformation in state timeline prepareTimelineFields
